requirements.txt
azure-ai-textanalytics==5.3.0b1
azure-common==1.1.28
azure-core==1.26.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msrest==0.7.1
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
azure-cognitiveservices-vision-computervision==0.7.0